requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0

# Database
sqlalchemy>=2.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Gmail API
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0

# API
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0

# HTTP client
httpx>=0.24.0

# Utilities
python-dateutil>=2.8.0

# Dev/Test
pytest>=7.4.0
pytest-asyncio>=0.21.0
